Inline list_issues, submit, report_failure into loop.py
Move tool implementations from tools.py into their call sites in
the tool provider. tools.py now only contains Pydantic argument
and result models.
